+++ analysis/tests/dummy_tool.ps1
@@ -2,6 +2,6 @@ Write-Output "Dummy tool!"
 Start-Sleep -Seconds 5
 Write-Output ("Your argument: " + $args[0])
 Write-Output ("SUT: " + $args[1])
-Write-Output "test file" > file1
+Set-Content -Path ./file1 -Value 'test file'
 mkdir -Force dir1 > $null
-echo "test file" > dir1/file2
+Set-Content -Path ./dir1/file2 -Value 'test file'